+++ examples/03-dev/demo/project/src/js/test/index.js
@@ -1,3 +1,8 @@
+/*!
+ * preserve comments
+ */
+
+
 
 require('bootstrap/dist/css/bootstrap.css');
 require('../../css/test/index.css');