+++ examples/03-dev/dist-net-all.js
@@ -10,4 +10,14 @@ projectConfig.network[100] = {
     backupHtml: !0
 };
 
+projectConfig.buildOptions = [];
+projectConfig.buildOptions[100] = {
+    renameHtmlMap: {
+        'test/index': 'parent/inner/index',
+        'test-2/index': 'parent/inner/index-2',
+        'test-3/index': 'parent/inner/index-3',
+        'test-5/index': 'parent/inner/index-5'
+    }
+};
+
 require('../../util/exec')('lila dist * -e 100');